+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- LightlyTrain now infers the best number of epochs based on the chosen method, dataset size and batch size.
+
 ### Deprecated
 
 ### Removed